refactor: refactor Kakao service and move API logic to external service module
