+++ backend/chidic-app/src/main/kotlin/com/chidicapp/security/auth/model/OAuth2UserDetails.kt
@@ -1,6 +1,6 @@
 package com.chidicapp.security.auth.model
 
-import com.chidicdomain.enum.Role
+import com.chidicdomain.type.Role
 import org.springframework.security.core.GrantedAuthority
 import org.springframework.security.core.authority.SimpleGrantedAuthority
 import org.springframework.security.oauth2.core.user.OAuth2User
