+++ backend/chidic-domain/src/main/kotlin/com/chidicdomain/domain/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package com.chidicdomain.domain.repository
 
-import com.chidicdomain.enum.Provider
+import com.chidicdomain.type.Provider
 import com.chidicdomain.domain.entity.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
