+++ backend/chidic-domain/src/main/kotlin/com/chidicdomain/type/Category.kt
@@ -1,4 +1,4 @@
-package com.chidicdomain.enum
+package com.chidicdomain.type
 
 enum class Category {
     OS,
