+++ backend/chidic-domain/src/main/kotlin/com/chidicdomain/type/FriendStatus.kt
@@ -1,4 +1,4 @@
-package com.chidicdomain.enum
+package com.chidicdomain.type
 
 enum class FriendStatus {
     PENDING,
