+++ backend/chidic-domain/src/main/kotlin/com/chidicdomain/type/Provider.kt
@@ -1,4 +1,4 @@
-package com.chidicdomain.enum
+package com.chidicdomain.type
 
 enum class Provider {
     NAVER,
